# dialect/sql: minor improvements to insert command