```
benchmark                                     old ns/op     new ns/op     delta
BenchmarkInsertBuilder_Default/default-16     6210          922           -85.15%

benchmark                                     old allocs     new allocs     delta
BenchmarkInsertBuilder_Default/default-16     86             16             -81.40%

benchmark                                     old bytes     new bytes     delta
BenchmarkInsertBuilder_Default/default-16     4098          1088          -73.45%

benchmark                                    old ns/op     new ns/op     delta
BenchmarkInsertBuilder_Small/sqlite3-16      2169          1801          -16.97%
BenchmarkInsertBuilder_Small/mysql-16        2140          1825          -14.72%
BenchmarkInsertBuilder_Small/postgres-16     2613          2223          -14.93%

benchmark                                    old allocs     new allocs     delta
BenchmarkInsertBuilder_Small/sqlite3-16      27             21             -22.22%
BenchmarkInsertBuilder_Small/mysql-16        27             21             -22.22%
BenchmarkInsertBuilder_Small/postgres-16     36             30             -16.67%

benchmark                                    old bytes     new bytes     delta
BenchmarkInsertBuilder_Small/sqlite3-16      1944          1304          -32.92%
BenchmarkInsertBuilder_Small/mysql-16        1944          1304          -32.92%
BenchmarkInsertBuilder_Small/postgres-16     1976          1336          -32.39%
```